fix: get rid of internal url parsing inside data providers (2 times)
