+++ src/providers/SearchProvider/index.ts
@@ -1,0 +1,2 @@
+export { default, useSearch } from './SearchProvider'
+export * from './types'
